[Camera Animations v2] Fix optimize bearing (#218)
* Fix optimize bearing

* revert change to example

* wip

* bring back presentsWithTransaction

* Cleanup

* cleanup debug app
