+++ Apps/DebugApp/DebugApp/DebugViewController.swift
@@ -79,9 +79,6 @@ public class DebugViewController: UIViewController {
          */
         mapView.on(.mapLoaded) { (event) in
             print("The map has finished loading... Event = \(event)")
-
-            let initialCenter = CLLocationCoordinate2D(latitude: 39.01305735102963, longitude: -77.01570412528032)
-            self.mapView.cameraManager.setCamera(centerCoordinate: initialCenter, zoom: 12)
         }
 
         /**
