+++ Sources/MapboxMaps/MapView/Configuration/RenderOptions.swift
@@ -22,4 +22,15 @@ public struct RenderOptions: Equatable {
     ///
     ///  The default value of this property is `true`.
     public var prefetchesTiles: Bool = true
+
+    /// A Boolean value that indicates whether the underlying `CAMetalLayer` of the `MapView`
+    /// presents its content using a CoreAnimation transaction
+    ///
+    /// By default, this is `false` resulting in the output of a rendering pass being displayed on
+    /// the `CAMetalLayer` as quickly as possible (and asynchronously). This typically results
+    /// in the fastest rendering performance.
+    ///
+    /// If, however, the `MapView` is overlaid with a `UIKit` element which must be pinned to a
+    /// particular lat-long, then setting this to `true` will result in better synchronization and less jitter.
+    public var presentsWithTransaction: Bool = true
 }
